Crate google-adsense1d4

This documentation was generated from AdSense crate version 0.0.1. The original source code can be found on github.

Features

Handle the following Resources with ease from the central hub ...

accounts (adclients.list, adunits.customchannels.list, adunits.get, adunits.get ad code, adunits.list, alerts.delete, alerts.list, customchannels.adunits.list, customchannels.get, customchannels.list, get, list, payments.list, reports.generate, reports.saved.generate, reports.saved.list, savedadstyles.get, savedadstyles.list and urlchannels.list)

adclients (list)

adunits (customchannels.list, get, get ad code and list)

alerts (delete and list)

customchannels (adunits.list, get and list)

metadata (dimensions.list and metrics.list)

payments (list)

reports (generate, saved.generate and saved.list)

savedadstyles (get and list)

urlchannels (list)

Download supported by ...

generate reports

reports.generate accounts

Everything else about the AdSense v1d4 API can be found at the official documentation site.

Not what you are looking for ? Find all other google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.accounts().adunits_list(...).doit()
let r = hub.accounts().customchannels_list(...).doit()
let r = hub.accounts().adunits_customchannels_list(...).doit()
let r = hub.accounts().alerts_delete(...).doit()
let r = hub.accounts().savedadstyles_get(...).doit()
let r = hub.accounts().reports_generate(...).doit()
let r = hub.accounts().list(...).doit()
let r = hub.accounts().adclients_list(...).doit()
let r = hub.accounts().reports_saved_generate(...).doit()
let r = hub.accounts().adunits_get(...).doit()
let r = hub.accounts().alerts_list(...).doit()
let r = hub.accounts().savedadstyles_list(...).doit()
let r = hub.accounts().customchannels_get(...).doit()
let r = hub.accounts().get(...).doit()
let r = hub.accounts().customchannels_adunits_list(...).doit()
let r = hub.accounts().payments_list(...).doit()
let r = hub.accounts().adunits_get_ad_code(...).doit()
let r = hub.accounts().reports_saved_list(...).doit()
let r = hub.accounts().urlchannels_list(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-adsense1d4 = "0.0.1"

A complete example

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate "google-adsense1d4" as adsense1d4;
use adsense1d4::cmn::Result;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use adsense1d4::AdSense;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = AdSense::new(hyper::Client::new(), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.accounts().reports_generate("accountId", "startDate", "endDate")
             .use_timezone_reporting(true)
             .start_index(-61)
             .add_sort("sadipscing")
             .add_metric("aliquyam")
             .max_results(-66)
             .locale("no")
             .add_filter("justo")
             .add_dimension("justo")
             .currency("et")
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::MissingAPIKey => println!("Missing API Key"),
    Result::MissingToken => println!("Missing Token"),
    Result::Failure(_) => println!("General Failure (Response doesn't print)"),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::Success(_) => println!("Success (value doesn't print)"),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

About Uploads and Downlods

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

About Customization/Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

About Parts

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifyable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

About Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are borrowed

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Modules

cmn

Structs

Account
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

adunits list accounts (none)

customchannels list accounts (none)

adunits customchannels list accounts (none)

alerts delete accounts (none)

savedadstyles get accounts (none)

reports generate accounts (none)

list accounts (none)

adclients list accounts (none)

reports saved generate accounts (none)

adunits get accounts (none)

alerts list accounts (none)

savedadstyles list accounts (none)

customchannels get accounts (none)

get accounts (response)

customchannels adunits list accounts (none)

payments list accounts (none)

adunits get ad code accounts (none)

reports saved list accounts (none)

urlchannels list accounts (none)

AccountAdclientListMethodBuilder
List all ad clients in the specified account.

AccountAdunitCustomchannelListMethodBuilder
List all custom channels which the specified ad unit belongs to.

AccountAdunitGetAdCodeMethodBuilder
Get ad code for the specified ad unit.

AccountAdunitGetMethodBuilder
Gets the specified ad unit in the specified ad client for the specified account.

AccountAdunitListMethodBuilder
List all ad units in the specified ad client for the specified account.

AccountAlertDeleteMethodBuilder
Dismiss (delete) the specified alert from the specified publisher AdSense account.

AccountAlertListMethodBuilder
List the alerts for the specified AdSense account.

AccountCustomchannelAdunitListMethodBuilder
List all ad units in the specified custom channel.

AccountCustomchannelGetMethodBuilder
Get the specified custom channel from the specified ad client for the specified account.

AccountCustomchannelListMethodBuilder
List all custom channels in the specified ad client for the specified account.

AccountGetMethodBuilder
Get information about the selected AdSense account.

AccountListMethodBuilder
List all accounts available to this AdSense account.

AccountMethodsBuilder
A builder providing access to all methods supported on account resources. It is not used directly, but through the AdSense hub.

AccountPaymentListMethodBuilder
List the payments for the specified AdSense account.

AccountReportGenerateMethodBuilder
Generate an AdSense report based on the report request sent in the query parameters. Returns the result as JSON; to retrieve output in CSV format specify "alt=csv" as a query parameter.

AccountReportSavedGenerateMethodBuilder
Generate an AdSense report based on the saved report ID sent in the query parameters.

AccountReportSavedListMethodBuilder
List all saved reports in the specified AdSense account.

AccountSavedadstyleGetMethodBuilder
List a specific saved ad style for the specified account.

AccountSavedadstyleListMethodBuilder
List all saved ad styles in the specified account.

AccountUrlchannelListMethodBuilder
List all URL channels in the specified ad client for the specified account.

Accounts
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list accounts (response)

AdClient
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

AdClients
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list adclients (response)

adclients list accounts (response)

AdCode
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

adunits get ad code accounts (response)

get ad code adunits (response)

AdSense
Central instance to access all AdSense related resource activities

AdStyle
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

AdStyleColors
The colors which are included in the style. These are represented as six hexadecimal characters, similar to HTML color codes, but without the leading hash.

This type is not used in any activity, and only used as part of another schema.

AdStyleFont
The font which is included in the style.

This type is not used in any activity, and only used as part of another schema.

AdUnit
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get adunits (response)

adunits get accounts (response)

AdUnitContentAdsSettings
Settings specific to content ads (AFC) and highend mobile content ads (AFMC).

This type is not used in any activity, and only used as part of another schema.

AdUnitContentAdsSettingsBackupOption
The backup option to be used in instances where no ad is available.

This type is not used in any activity, and only used as part of another schema.

AdUnitFeedAdsSettings
Settings specific to feed ads (AFF).

This type is not used in any activity, and only used as part of another schema.

AdUnitMobileContentAdsSettings
Settings specific to WAP mobile content ads (AFMC).

This type is not used in any activity, and only used as part of another schema.

AdUnits
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

adunits list accounts (response)

list adunits (response)

customchannels adunits list accounts (response)

adunits list customchannels (response)

AdclientListMethodBuilder
List all ad clients in this AdSense account.

AdclientMethodsBuilder
A builder providing access to all methods supported on adclient resources. It is not used directly, but through the AdSense hub.

AdsenseReportsGenerateResponse
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

saved generate reports (response)

reports generate accounts (response)

reports saved generate accounts (response)

generate reports (response)

AdsenseReportsGenerateResponseHeaders
The header information of the columns requested in the report. This is a list of headers; one for each dimension in the request, followed by one for each metric in the request.

This type is not used in any activity, and only used as part of another schema.

AdunitCustomchannelListMethodBuilder
List all custom channels which the specified ad unit belongs to.

AdunitGetAdCodeMethodBuilder
Get ad code for the specified ad unit.

AdunitGetMethodBuilder
Gets the specified ad unit in the specified ad client.

AdunitListMethodBuilder
List all ad units in the specified ad client for this AdSense account.

AdunitMethodsBuilder
A builder providing access to all methods supported on adunit resources. It is not used directly, but through the AdSense hub.

Alert
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list alerts (none)

delete alerts (none)

AlertDeleteMethodBuilder
Dismiss (delete) the specified alert from the publisher's AdSense account.

AlertListMethodBuilder
List the alerts for this AdSense account.

AlertMethodsBuilder
A builder providing access to all methods supported on alert resources. It is not used directly, but through the AdSense hub.

Alerts
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

alerts list accounts (response)

list alerts (response)

CustomChannel
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

customchannels get accounts (response)

get customchannels (response)

CustomChannelTargetingInfo
The targeting information of this custom channel, if activated.

This type is not used in any activity, and only used as part of another schema.

CustomChannels
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

customchannels list accounts (response)

customchannels list adunits (response)

adunits customchannels list accounts (response)

list customchannels (response)

CustomchannelAdunitListMethodBuilder
List all ad units in the specified custom channel.

CustomchannelGetMethodBuilder
Get the specified custom channel from the specified ad client.

CustomchannelListMethodBuilder
List all custom channels in the specified ad client for this AdSense account.

CustomchannelMethodsBuilder
A builder providing access to all methods supported on customchannel resources. It is not used directly, but through the AdSense hub.

Metadata
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

metrics list metadata (response)

dimensions list metadata (response)

MetadataDimensionListMethodBuilder
List the metadata for the dimensions available to this AdSense account.

MetadataMethodsBuilder
A builder providing access to all methods supported on metadata resources. It is not used directly, but through the AdSense hub.

MetadataMetricListMethodBuilder
List the metadata for the metrics available to this AdSense account.

Payment
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list payments (none)

PaymentListMethodBuilder
List the payments for this AdSense account.

PaymentMethodsBuilder
A builder providing access to all methods supported on payment resources. It is not used directly, but through the AdSense hub.

Payments
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

payments list accounts (response)

list payments (response)

ReportGenerateMethodBuilder
Generate an AdSense report based on the report request sent in the query parameters. Returns the result as JSON; to retrieve output in CSV format specify "alt=csv" as a query parameter.

ReportMethodsBuilder
A builder providing access to all methods supported on report resources. It is not used directly, but through the AdSense hub.

ReportSavedGenerateMethodBuilder
Generate an AdSense report based on the saved report ID sent in the query parameters.

ReportSavedListMethodBuilder
List all saved reports in this AdSense account.

ReportingMetadataEntry
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

SavedAdStyle
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get savedadstyles (response)

savedadstyles get accounts (response)

SavedAdStyles
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

savedadstyles list accounts (response)

list savedadstyles (response)

SavedReport
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

SavedReports
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

saved list reports (response)

reports saved list accounts (response)

SavedadstyleGetMethodBuilder
Get a specific saved ad style from the user's account.

SavedadstyleListMethodBuilder
List all saved ad styles in the user's account.

SavedadstyleMethodsBuilder
A builder providing access to all methods supported on savedadstyle resources. It is not used directly, but through the AdSense hub.

UrlChannel
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

UrlChannels
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list urlchannels (response)

urlchannels list accounts (response)

UrlchannelListMethodBuilder
List all URL channels in the specified ad client for this AdSense account.

UrlchannelMethodsBuilder
A builder providing access to all methods supported on urlchannel resources. It is not used directly, but through the AdSense hub.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.