Struct google_adsense1d4::Account

pub struct Account {
    pub kind: Option<String>,
    pub premium: Option<bool>,
    pub name: Option<String>,
    pub timezone: Option<String>,
    pub creation_time: Option<String>,
    pub id: Option<String>,
    pub sub_accounts: Option<Vec<Account>>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

adunits list accounts (none)

customchannels list accounts (none)

adunits customchannels list accounts (none)

alerts delete accounts (none)

savedadstyles get accounts (none)

reports generate accounts (none)

list accounts (none)

adclients list accounts (none)

reports saved generate accounts (none)

adunits get accounts (none)

alerts list accounts (none)

savedadstyles list accounts (none)

customchannels get accounts (none)

get accounts (response)

customchannels adunits list accounts (none)

payments list accounts (none)

adunits get ad code accounts (none)

reports saved list accounts (none)

urlchannels list accounts (none)

Fields
kind: Option<String>

Kind of resource this is, in this case adsense#account.

premium: Option<bool>

Whether this account is premium.

name: Option<String>

Name of this account.

timezone: Option<String>

AdSense timezone of this account.

creation_time: Option<String>

no description provided

id: Option<String>

Unique identifier of this account.

sub_accounts: Option<Vec<Account>>

Sub accounts of the this account.