Struct google_adsense1d4::AccountCustomchannelGetCall

pub struct AccountCustomchannelGetCall<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

Get the specified custom channel from the specified ad client for the specified account.

A builder for the customchannels.get method supported by a account resource. It is not used directly, but through a AccountMethods instance.

Example

Instantiate a resource method builder

 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.accounts().customchannels_get("accountId", "adClientId", "customChannelId")
             .doit();

Methods

impl<'a, C, A> AccountCustomchannelGetCall<'a, C, A> where C: BorrowMut<Client>, A: GetToken

fn doit(self) -> Result<(Response, CustomChannel)>

Perform the operation you have build so far.

fn account_id(self, new_value: &str) -> AccountCustomchannelGetCall<'a, C, A>

Account to which the ad client belongs.

Sets the account id path property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

fn ad_client_id(self, new_value: &str) -> AccountCustomchannelGetCall<'a, C, A>

Ad client which contains the custom channel.

Sets the ad client id path property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

fn custom_channel_id(self, new_value: &str) -> AccountCustomchannelGetCall<'a, C, A>

Custom channel to retrieve.

Sets the custom channel id path property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

fn delegate(self, new_value: &'a mut Delegate) -> AccountCustomchannelGetCall<'a, C, A>

The delegate implementation is consulted whenever there is an intermediate result, or if something goes wrong while executing the actual API request.

It should be used to handle progress information, and to implement a certain level of resilience.

Sets the delegate property to the given value.

fn param<T>(self, name: T, value: T) -> AccountCustomchannelGetCall<'a, C, A> where T: AsRef<str>

Set any additional parameter of the query string used in the request. It should be used to set parameters which are not yet available through their own setters.

Please note that this method must not be used to set any of the known paramters which have their own setter method. If done anyway, the request will fail.

Additional Parameters

quotaUser (query-string) - Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters. Overrides userIp if both are provided.

oauth_token (query-string) - OAuth 2.0 token for the current user.

key (query-string) - API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

prettyPrint (query-boolean) - Returns response with indentations and line breaks.

userIp (query-string) - IP address of the site where the request originates. Use this if you want to enforce per-user limits.

fields (query-string) - Selector specifying which fields to include in a partial response.

alt (query-string) - Data format for the response.

fn add_scope<T>(self, scope: T) -> AccountCustomchannelGetCall<'a, C, A> where T: AsRef<str>

Identifies the authorization scope for the method you are building.

Use this method to actively specify which scope should be used, instead the default Scope variant Scope::Readonly.

The scope will be added to a set of scopes. This is important as one can maintain access tokens for more than one scope.

Usually there is more than one suitable scope to authorize an operation, some of which may encompass more rights than others. For example, for listing resources, a read-only scope will be sufficient, a read-write scope will do as well.