Struct google_adsense1d4::AdUnit

pub struct AdUnit {
    pub status: Option<String>,
    pub kind: Option<String>,
    pub code: Option<String>,
    pub name: Option<String>,
    pub feed_ads_settings: Option<AdUnitFeedAdsSettings>,
    pub saved_style_id: Option<String>,
    pub content_ads_settings: Option<AdUnitContentAdsSettings>,
    pub id: Option<String>,
    pub mobile_content_ads_settings: Option<AdUnitMobileContentAdsSettings>,
    pub custom_style: Option<AdStyle>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get adunits (response)

adunits get accounts (response)

Fields

status

Status of this ad unit. Possible values are: NEW: Indicates that the ad unit was created within the last seven days and does not yet have any activity associated with it.

ACTIVE: Indicates that there has been activity on this ad unit in the last seven days.

INACTIVE: Indicates that there has been no activity on this ad unit in the last seven days.

kind

Kind of resource this is, in this case adsense#adUnit.

code

Identity code of this ad unit, not necessarily unique across ad clients.

name

Name of this ad unit.

feed_ads_settings

Settings specific to feed ads (AFF) - deprecated.

saved_style_id

ID of the saved ad style which holds this ad unit's style information.

content_ads_settings

Settings specific to content ads (AFC) and highend mobile content ads (AFMC - deprecated).

id

Unique identifier of this ad unit. This should be considered an opaque identifier; it is not safe to rely on it being in any particular format.

mobile_content_ads_settings

Settings specific to WAP mobile content ads (AFMC) - deprecated.

custom_style

Custom style information specific to this ad unit.