Struct google_adsense1d4::Alert

pub struct Alert {
    pub kind: Option<String>,
    pub severity: Option<String>,
    pub is_dismissible: Option<bool>,
    pub message: Option<String>,
    pub type_: Option<String>,
    pub id: Option<String>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list alerts (none)

delete alerts (none)

Fields

kind

Kind of resource this is, in this case adsense#alert.

severity

Severity of this alert. Possible values: INFO, WARNING, SEVERE.

is_dismissible

Whether this alert can be dismissed.

message

The localized alert message.

type_

Type of this alert. Possible values: SELF_HOLD, MIGRATED_TO_BILLING3, ADDRESS_PIN_VERIFICATION, PHONE_PIN_VERIFICATION, CORPORATE_ENTITY, GRAYLISTED_PUBLISHER, API_HOLD.

id

Unique identifier of this alert. This should be considered an opaque identifier; it is not safe to rely on it being in any particular format.