Struct google_adsense1d4::MetadataMethods

pub struct MetadataMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on metadata resources. It is not used directly, but through the AdSense hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_adsense1d4 as adsense1d4;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use adsense1d4::AdSense;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = AdSense::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `dimensions_list(...)` and `metrics_list(...)`
// to build up your call.
let rb = hub.metadata();

Methods

impl<'a, C, A> MetadataMethods<'a, C, A>

fn dimensions_list(&self) -> MetadataDimensionListCall<'a, C, A>

Create a builder to help you perform the following task:

List the metadata for the dimensions available to this AdSense account.

fn metrics_list(&self) -> MetadataMetricListCall<'a, C, A>

Create a builder to help you perform the following task:

List the metadata for the metrics available to this AdSense account.