Struct google_adsense1d4::ReportMethods

pub struct ReportMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on report resources. It is not used directly, but through the AdSense hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_adsense1d4 as adsense1d4;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use adsense1d4::AdSense;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = AdSense::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `generate(...)`, `saved_generate(...)` and `saved_list(...)`
// to build up your call.
let rb = hub.reports();

Methods

impl<'a, C, A> ReportMethods<'a, C, A>

fn saved_list(&self) -> ReportSavedListCall<'a, C, A>

Create a builder to help you perform the following task:

List all saved reports in this AdSense account.

fn generate(&self, start_date: &str, end_date: &str) -> ReportGenerateCall<'a, C, A>

Create a builder to help you perform the following task:

Generate an AdSense report based on the report request sent in the query parameters. Returns the result as JSON; to retrieve output in CSV format specify "alt=csv" as a query parameter.

Arguments

startDate - Start of the date range to report on in "YYYY-MM-DD" format, inclusive.

endDate - End of the date range to report on in "YYYY-MM-DD" format, inclusive.

fn saved_generate(&self, saved_report_id: &str) -> ReportSavedGenerateCall<'a, C, A>

Create a builder to help you perform the following task:

Generate an AdSense report based on the saved report ID sent in the query parameters.

Arguments

savedReportId - The saved report to retrieve.