Crate google_adsense1d4

This documentation was generated from AdSense crate version 1.0.9+20190703, where 20190703 is the exact revision of the adsense:v1.4 schema built by the mako code generator v1.0.9.

Everything else about the AdSense v1d4 API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

accounts

adclients get ad code, adclients list, adunits customchannels list, adunits get, adunits get ad code, adunits list, alerts delete, alerts list, customchannels adunits list, customchannels get, customchannels list, get, list, payments list, reports generate, reports saved generate, reports saved list, savedadstyles get, savedadstyles list and urlchannels list

adclients

list

adunits

customchannels list, get, get ad code and list

alerts

delete and list

customchannels

adunits list, get and list

metadata

dimensions list and metrics list

payments

list

reports

generate, saved generate and saved list

savedadstyles

get and list

urlchannels

list

Download supported by ...

generate reports

reports generate accounts

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.accounts().adunits_list(...).doit()
let r = hub.accounts().customchannels_list(...).doit()
let r = hub.accounts().adunits_customchannels_list(...).doit()
let r = hub.accounts().alerts_delete(...).doit()
let r = hub.accounts().savedadstyles_get(...).doit()
let r = hub.accounts().reports_generate(...).doit()
let r = hub.accounts().list(...).doit()
let r = hub.accounts().adclients_list(...).doit()
let r = hub.accounts().reports_saved_generate(...).doit()
let r = hub.accounts().adunits_get(...).doit()
let r = hub.accounts().alerts_list(...).doit()
let r = hub.accounts().savedadstyles_list(...).doit()
let r = hub.accounts().customchannels_get(...).doit()
let r = hub.accounts().get(...).doit()
let r = hub.accounts().customchannels_adunits_list(...).doit()
let r = hub.accounts().payments_list(...).doit()
let r = hub.accounts().adunits_get_ad_code(...).doit()
let r = hub.accounts().adclients_get_ad_code(...).doit()
let r = hub.accounts().reports_saved_list(...).doit()
let r = hub.accounts().urlchannels_list(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-adsense1d4 = "*"
# This project intentionally uses an old version of Hyper. See
# https://github.com/Byron/google-apis-rs/issues/173 for more
# information.
hyper = "^0.10"
hyper-rustls = "^0.6"
serde = "^1.0"
serde_json = "^1.0"
yup-oauth2 = "^1.0"

A complete example

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_adsense1d4 as adsense1d4;
use adsense1d4::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use adsense1d4::AdSense;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = AdSense::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.accounts().reports_generate("accountId", "startDate", "endDate")
             .use_timezone_reporting(true)
             .start_index(-61)
             .add_sort("sadipscing")
             .add_metric("aliquyam")
             .max_results(-66)
             .locale("no")
             .add_filter("justo")
             .add_dimension("justo")
             .currency("et")
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

Account
There is no detailed description.

AccountAdclientGetAdCodeCall
Get Auto ad code for a given ad client.

AccountAdclientListCall
List all ad clients in the specified account.

AccountAdunitCustomchannelListCall
List all custom channels which the specified ad unit belongs to.

AccountAdunitGetAdCodeCall
Get ad code for the specified ad unit.

AccountAdunitGetCall
Gets the specified ad unit in the specified ad client for the specified account.

AccountAdunitListCall
List all ad units in the specified ad client for the specified account.

AccountAlertDeleteCall
Dismiss (delete) the specified alert from the specified publisher AdSense account.

AccountAlertListCall
List the alerts for the specified AdSense account.

AccountCustomchannelAdunitListCall
List all ad units in the specified custom channel.

AccountCustomchannelGetCall
Get the specified custom channel from the specified ad client for the specified account.

AccountCustomchannelListCall
List all custom channels in the specified ad client for the specified account.

AccountGetCall
Get information about the selected AdSense account.

AccountListCall
List all accounts available to this AdSense account.

AccountMethods
A builder providing access to all methods supported on account resources. It is not used directly, but through the AdSense hub.

AccountPaymentListCall
List the payments for the specified AdSense account.

AccountReportGenerateCall
Generate an AdSense report based on the report request sent in the query parameters. Returns the result as JSON; to retrieve output in CSV format specify "alt=csv" as a query parameter.

AccountReportSavedGenerateCall
Generate an AdSense report based on the saved report ID sent in the query parameters.

AccountReportSavedListCall
List all saved reports in the specified AdSense account.

AccountSavedadstyleGetCall
List a specific saved ad style for the specified account.

AccountSavedadstyleListCall
List all saved ad styles in the specified account.

AccountUrlchannelListCall
List all URL channels in the specified ad client for the specified account.

Accounts
There is no detailed description.

AdClient
There is no detailed description.

AdClients
There is no detailed description.

AdCode
There is no detailed description.

AdSense
Central instance to access all AdSense related resource activities

AdStyle
There is no detailed description.

AdStyleColors
The colors which are included in the style. These are represented as six hexadecimal characters, similar to HTML color codes, but without the leading hash.

AdStyleFont
The font which is included in the style.

AdUnit
There is no detailed description.

AdUnitContentAdsSettings
Settings specific to content ads (AFC) and highend mobile content ads (AFMC - deprecated).

AdUnitContentAdsSettingsBackupOption
The backup option to be used in instances where no ad is available.

AdUnitFeedAdsSettings
Settings specific to feed ads (AFF) - deprecated.

AdUnitMobileContentAdsSettings
Settings specific to WAP mobile content ads (AFMC) - deprecated.

AdUnits
There is no detailed description.

AdclientListCall
List all ad clients in this AdSense account.

AdclientMethods
A builder providing access to all methods supported on adclient resources. It is not used directly, but through the AdSense hub.

AdsenseReportsGenerateResponse
There is no detailed description.

AdsenseReportsGenerateResponseHeaders
The header information of the columns requested in the report. This is a list of headers; one for each dimension in the request, followed by one for each metric in the request.

AdunitCustomchannelListCall
List all custom channels which the specified ad unit belongs to.

AdunitGetAdCodeCall
Get ad code for the specified ad unit.

AdunitGetCall
Gets the specified ad unit in the specified ad client.

AdunitListCall
List all ad units in the specified ad client for this AdSense account.

AdunitMethods
A builder providing access to all methods supported on adunit resources. It is not used directly, but through the AdSense hub.

Alert
There is no detailed description.

AlertDeleteCall
Dismiss (delete) the specified alert from the publisher's AdSense account.

AlertListCall
List the alerts for this AdSense account.

AlertMethods
A builder providing access to all methods supported on alert resources. It is not used directly, but through the AdSense hub.

Alerts
There is no detailed description.

Chunk

ContentRange
Implements the Content-Range header, for serialization only

CustomChannel
There is no detailed description.

CustomChannelTargetingInfo
The targeting information of this custom channel, if activated.

CustomChannels
There is no detailed description.

CustomchannelAdunitListCall
List all ad units in the specified custom channel.

CustomchannelGetCall
Get the specified custom channel from the specified ad client.

CustomchannelListCall
List all custom channels in the specified ad client for this AdSense account.

CustomchannelMethods
A builder providing access to all methods supported on customchannel resources. It is not used directly, but through the AdSense hub.

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

DummyNetworkStream

ErrorResponse
A utility to represent detailed errors we might see in case there are BadRequests. The latter happen if the sent parameters or request structures are unsound

JsonServerError
A utility type which can decode a server response that indicates error

Metadata
There is no detailed description.

MetadataDimensionListCall
List the metadata for the dimensions available to this AdSense account.

MetadataMethods
A builder providing access to all methods supported on metadata resources. It is not used directly, but through the AdSense hub.

MetadataMetricListCall
List the metadata for the metrics available to this AdSense account.

MethodInfo
Contains information about an API request.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

Payment
There is no detailed description.

PaymentListCall
List the payments for this AdSense account.

PaymentMethods
A builder providing access to all methods supported on payment resources. It is not used directly, but through the AdSense hub.

Payments
There is no detailed description.

RangeResponseHeader

ReportGenerateCall
Generate an AdSense report based on the report request sent in the query parameters. Returns the result as JSON; to retrieve output in CSV format specify "alt=csv" as a query parameter.

ReportMethods
A builder providing access to all methods supported on report resources. It is not used directly, but through the AdSense hub.

ReportSavedGenerateCall
Generate an AdSense report based on the saved report ID sent in the query parameters.

ReportSavedListCall
List all saved reports in this AdSense account.

ReportingMetadataEntry
There is no detailed description.

ResumableUploadHelper
A utility type to perform a resumable upload from start to end.

SavedAdStyle
There is no detailed description.

SavedAdStyles
There is no detailed description.

SavedReport
There is no detailed description.

SavedReports
There is no detailed description.

SavedadstyleGetCall
Get a specific saved ad style from the user's account.

SavedadstyleListCall
List all saved ad styles in the user's account.

SavedadstyleMethods
A builder providing access to all methods supported on savedadstyle resources. It is not used directly, but through the AdSense hub.

ServerError

ServerMessage

UrlChannel
There is no detailed description.

UrlChannels
There is no detailed description.

UrlchannelListCall
List all URL channels in the specified ad client for this AdSense account.

UrlchannelMethods
A builder providing access to all methods supported on urlchannel resources. It is not used directly, but through the AdSense hub.

XUploadContentType
The X-Upload-Content-Type header.

Enums

Error

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

MethodsBuilder
Identifies types for building methods of a particular resource type

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

Resource
Identifies types which can be inserted and deleted. Types with this trait are most commonly used by clients of this API.

ResponseResult
Identifies types which are used in API responses.

ToParts
A trait for all types that can convert themselves into a parts string

UnusedType
Identifies types which are not actually used by the API This might be a bug within the google API schema.

Functions

remove_json_null_values

Type Definitions

Result
A universal result type used as return for all calls.