Struct google_adsense1d4::AccountMethods

pub struct AccountMethods<'a, C, A> where
    C: 'a,
    A: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on account resources. It is not used directly, but through the AdSense hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_adsense1d4 as adsense1d4;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use adsense1d4::AdSense;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = AdSense::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `adclients_get_ad_code(...)`, `adclients_list(...)`, `adunits_customchannels_list(...)`, `adunits_get(...)`, `adunits_get_ad_code(...)`, `adunits_list(...)`, `alerts_delete(...)`, `alerts_list(...)`, `customchannels_adunits_list(...)`, `customchannels_get(...)`, `customchannels_list(...)`, `get(...)`, `list(...)`, `payments_list(...)`, `reports_generate(...)`, `reports_saved_generate(...)`, `reports_saved_list(...)`, `savedadstyles_get(...)`, `savedadstyles_list(...)` and `urlchannels_list(...)`
// to build up your call.
let rb = hub.accounts();

Methods

impl<'a, C, A> AccountMethods<'a, C, A>

pub fn alerts_delete(
&self,
account_id: &str,
alert_id: &str
) -> AccountAlertDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Dismiss (delete) the specified alert from the specified publisher AdSense account.

Arguments

accountId - Account which contains the ad unit.

alertId - Alert to delete.

pub fn alerts_list(&self, account_id: &str) -> AccountAlertListCall<'a, C, A>

Create a builder to help you perform the following task:

List the alerts for the specified AdSense account.

Arguments

accountId - Account for which to retrieve the alerts.

pub fn reports_saved_generate(
&self,
account_id: &str,
saved_report_id: &str
) -> AccountReportSavedGenerateCall<'a, C, A>

Create a builder to help you perform the following task:

Generate an AdSense report based on the saved report ID sent in the query parameters.

Arguments

accountId - Account to which the saved reports belong.

savedReportId - The saved report to retrieve.

pub fn list(&self) -> AccountListCall<'a, C, A>

Create a builder to help you perform the following task:

List all accounts available to this AdSense account.

pub fn payments_list(
&self,
account_id: &str
) -> AccountPaymentListCall<'a, C, A>

Create a builder to help you perform the following task:

List the payments for the specified AdSense account.

Arguments

accountId - Account for which to retrieve the payments.

pub fn savedadstyles_get(
&self,
account_id: &str,
saved_ad_style_id: &str
) -> AccountSavedadstyleGetCall<'a, C, A>

Create a builder to help you perform the following task:

List a specific saved ad style for the specified account.

Arguments

accountId - Account for which to get the saved ad style.

savedAdStyleId - Saved ad style to retrieve.

pub fn adunits_get(
&self,
account_id: &str,
ad_client_id: &str,
ad_unit_id: &str
) -> AccountAdunitGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the specified ad unit in the specified ad client for the specified account.

Arguments

accountId - Account to which the ad client belongs.

adClientId - Ad client for which to get the ad unit.

adUnitId - Ad unit to retrieve.

pub fn customchannels_get(
&self,
account_id: &str,
ad_client_id: &str,
custom_channel_id: &str
) -> AccountCustomchannelGetCall<'a, C, A>

Create a builder to help you perform the following task:

Get the specified custom channel from the specified ad client for the specified account.

Arguments

accountId - Account to which the ad client belongs.

adClientId - Ad client which contains the custom channel.

customChannelId - Custom channel to retrieve.

pub fn customchannels_adunits_list(
&self,
account_id: &str,
ad_client_id: &str,
custom_channel_id: &str
) -> AccountCustomchannelAdunitListCall<'a, C, A>

Create a builder to help you perform the following task:

List all ad units in the specified custom channel.

Arguments

accountId - Account to which the ad client belongs.

adClientId - Ad client which contains the custom channel.

customChannelId - Custom channel for which to list ad units.

pub fn reports_saved_list(
&self,
account_id: &str
) -> AccountReportSavedListCall<'a, C, A>

Create a builder to help you perform the following task:

List all saved reports in the specified AdSense account.

Arguments

accountId - Account to which the saved reports belong.

pub fn urlchannels_list(
&self,
account_id: &str,
ad_client_id: &str
) -> AccountUrlchannelListCall<'a, C, A>

Create a builder to help you perform the following task:

List all URL channels in the specified ad client for the specified account.

Arguments

accountId - Account to which the ad client belongs.

adClientId - Ad client for which to list URL channels.

pub fn adunits_list(
&self,
account_id: &str,
ad_client_id: &str
) -> AccountAdunitListCall<'a, C, A>

Create a builder to help you perform the following task:

List all ad units in the specified ad client for the specified account.

Arguments

accountId - Account to which the ad client belongs.

adClientId - Ad client for which to list ad units.

pub fn reports_generate(
&self,
account_id: &str,
start_date: &str,
end_date: &str
) -> AccountReportGenerateCall<'a, C, A>

Create a builder to help you perform the following task:

Generate an AdSense report based on the report request sent in the query parameters. Returns the result as JSON; to retrieve output in CSV format specify "alt=csv" as a query parameter.

Arguments

accountId - Account upon which to report.

startDate - Start of the date range to report on in "YYYY-MM-DD" format, inclusive.

endDate - End of the date range to report on in "YYYY-MM-DD" format, inclusive.

pub fn adclients_list(
&self,
account_id: &str
) -> AccountAdclientListCall<'a, C, A>

Create a builder to help you perform the following task:

List all ad clients in the specified account.

Arguments

accountId - Account for which to list ad clients.

pub fn customchannels_list(
&self,
account_id: &str,
ad_client_id: &str
) -> AccountCustomchannelListCall<'a, C, A>

Create a builder to help you perform the following task:

List all custom channels in the specified ad client for the specified account.

Arguments

accountId - Account to which the ad client belongs.

adClientId - Ad client for which to list custom channels.

pub fn savedadstyles_list(
&self,
account_id: &str
) -> AccountSavedadstyleListCall<'a, C, A>

Create a builder to help you perform the following task:

List all saved ad styles in the specified account.

Arguments

accountId - Account for which to list saved ad styles.

pub fn adunits_get_ad_code(
&self,
account_id: &str,
ad_client_id: &str,
ad_unit_id: &str
) -> AccountAdunitGetAdCodeCall<'a, C, A>

Create a builder to help you perform the following task:

Get ad code for the specified ad unit.

Arguments

accountId - Account which contains the ad client.

adClientId - Ad client with contains the ad unit.

adUnitId - Ad unit to get the code for.

pub fn adunits_customchannels_list(
&self,
account_id: &str,
ad_client_id: &str,
ad_unit_id: &str
) -> AccountAdunitCustomchannelListCall<'a, C, A>

Create a builder to help you perform the following task:

List all custom channels which the specified ad unit belongs to.

Arguments

accountId - Account to which the ad client belongs.

adClientId - Ad client which contains the ad unit.

adUnitId - Ad unit for which to list custom channels.

pub fn get(&self, account_id: &str) -> AccountGetCall<'a, C, A>

Create a builder to help you perform the following task:

Get information about the selected AdSense account.

Arguments

accountId - Account to get information about.

pub fn adclients_get_ad_code(
&self,
account_id: &str,
ad_client_id: &str
) -> AccountAdclientGetAdCodeCall<'a, C, A>

Create a builder to help you perform the following task:

Get Auto ad code for a given ad client.

Arguments

accountId - Account which contains the ad client.

adClientId - Ad client to get the code for.